Changes opacity by 10% on each hover
This completes the exercise and additional challenges.
I removed setAttribute as it kept resetting, so I used elem.style.X directly.

Added an If statement to check opacity and increment it by 0.1 until it reaches maximum of 1.0.

An additional div is used in order to keep the box. Opacity means the borders are also removed, I didn't want that.